Remove options and use localStorage for the Gmail.
This extension defaults to production but can be changed to point to any server.

Chrome does not allow localStorage to be shared between an extension and a tab, and `chrome.storage` (the recommended solution) is unreliable about issuing change events. This means that option-to-tab communication is broken. Tested in both Chrome and the current-as-of-today Chrome Canary.

Since the only people who would ever point to UAT are testers, and we can assume a minimum level of technical competence, The way to change the destination is to issue a localStorage.setItem. See README.md for mode details.